perf: 适配新版写法
