+++ packages/s2-core/src/group/grid-group.ts
@@ -7,7 +7,8 @@ import {
 import type { GridInfo } from '../common/interface';
 import type { GridGroupConstructorParameters } from '../common/interface/group';
 import type { SpreadSheet } from '../sheet-type/spread-sheet';
-import { batchSetStyle, renderLine } from '../utils/g-renders';
+import { renderLine } from '../utils/g-renders';
+import { batchSetStyle } from '../utils/g-utils';
 
 export class GridGroup extends Group {
   protected s2: SpreadSheet;
