+++ packages/s2-core/src/utils/g-renders.ts
@@ -17,7 +17,7 @@ import {
   type RectStyleProps,
   type TextStyleProps,
 } from '@antv/g';
-import { get, isArray, isEmpty, isFunction, set } from 'lodash';
+import { isArray, isEmpty, isFunction } from 'lodash';
 import { GuiIcon, type GuiIconCfg } from '../common/icons/gui-icon';
 import { CustomText } from '../engine/CustomText';
 
@@ -143,26 +143,4 @@ export function renderTreeIcon(options: {
   }
 
   return iconShape;
-}
-
-export function batchSetStyle<
-  T extends DisplayObject,
-  S extends BaseStyleProps,
->(obj: T, style: S) {
-  obj.setAttributes(style, { skipDispatchAttrModifiedEvent: true });
-}
-
-export function createOrUpdateRect(
-  propertyPath: string,
-  style: RectStyleProps,
-) {
-  // @ts-ignore
-  const context = this as any;
-  const obj = get(context, propertyPath);
-
-  if (!obj) {
-    set(context, propertyPath, new Rect({ style }));
-  } else {
-    batchSetStyle(obj, style);
-  }
 }
