+++ packages/s2-core/src/utils/g-utils.ts
@@ -1,13 +1,29 @@
-import { BaseStyleProps, DisplayObject } from '@antv/g';
+import {
+  BaseStyleProps,
+  DisplayObject,
+  Rect,
+  type RectStyleProps,
+} from '@antv/g';
+import { get, set } from 'lodash';
 
 export function batchSetStyle<
   T extends DisplayObject,
-  S extends BaseStyleProps & {
-    x?: number | string;
-    y?: number | string;
-  },
+  S extends BaseStyleProps,
 >(obj: T, style: S) {
-  for (const styleKey in style) {
-    obj.style[styleKey] = style[styleKey];
+  obj.setAttributes(style, { skipDispatchAttrModifiedEvent: true });
+}
+
+export function createOrUpdateRect(
+  propertyPath: string,
+  style: RectStyleProps,
+) {
+  // @ts-ignore
+  const context = this as any;
+  const obj = get(context, propertyPath);
+
+  if (!obj) {
+    set(context, propertyPath, new Rect({ style }));
+  } else {
+    batchSetStyle(obj, style);
   }
 }
